new config driver for json file
